refactor(boundaries): extract activity-feed text renderers to core protocol (refs #1363)
The cockpit's static-pane fallback (cockpit._build_cockpit_detail_dispatch
when kind == "activity") lazy-imported render_activity_feed_text from the
activity_feed plugin's feed_panel. That kept the optional plugin effectively
non-optional: disabling it would silently break the operator dashboard's
activity pane.

Hoist the four plain-text renderers (format_entry_row,
compute_project_column_width, render_entries_as_text, render_activity_feed_text)
plus the _project_label helper into pollypm.activity_feed_protocol next to
the existing format_relative_time / render_entry_detail helpers. The
config-in/text-out shim resolves the projector through
pollypm.activity_projector_registry.build_activity_projector so the cockpit
never reaches into plugins_builtin.

feed_panel.py re-exports the names so the plugin's own CLI + tests +
monkeypatch paths keep working unchanged. No behaviour change — the helpers
are copied verbatim and the renderer accepts the same FeedEntry shape
structurally via the ActivityFeedEntry Protocol.
