+++ src/pollypm/plugins_builtin/activity_feed/cockpit/feed_panel.py
@@ -25,7 +25,7 @@ import logging
 from dataclasses import dataclass, replace
 from datetime import UTC, datetime, timedelta
 from pathlib import Path
-from typing import Any, Iterable
+from typing import Iterable
 
 from pollypm.plugins_builtin.activity_feed.handlers.event_projector import (
     EventProjector,
@@ -167,120 +167,20 @@ _SEVERITY_COLOURS = {
 
 
 # Re-exported from the core shared module so callers in ``pollypm.cockpit*``
-# (and other peer plugins) can use the helper without importing through
-# ``plugins_builtin`` (#1363). The plugin keeps the name on its own surface
-# for back-compat with existing imports + monkeypatch paths.
-from pollypm.activity_feed_protocol import format_relative_time  # noqa: F401, E402
-
-
-def _project_label(entry: FeedEntry) -> str:
-    """Project key as displayed in the feed (``"-"`` for empty).
-
-    Centralised so the auto-fit width math and ``format_entry_row``
-    agree on the same string. Empty / ``None`` falls back to ``"-"``
-    so an unknown project still aligns in the column.
-    """
-    return entry.project or "-"
-
-
-def compute_project_column_width(
-    entries: Iterable[FeedEntry], *, minimum: int = 1,
-) -> int:
-    """Width of the widest project key in ``entries`` (auto-fit).
-
-    Used by :func:`format_entry_row` so the project column expands to
-    fit long keys like ``blackjack-trainer`` while keeping every row
-    aligned. Empty input falls back to ``minimum`` so a header-only
-    render still produces a sane column. See #929.
-    """
-    widest = minimum
-    for entry in entries:
-        widest = max(widest, len(_project_label(entry)))
-    return widest
-
-
-def format_entry_row(
-    entry: FeedEntry,
-    *,
-    now: datetime | None = None,
-    project_width: int | None = None,
-) -> str:
-    """Render one FeedEntry as a single plain-text row.
-
-    Layout:: ``[rel] [project] [actor] verb summary``. Severity is not
-    encoded here — the Textual renderer applies colour; the plain-text
-    renderer prefixes a ``!`` on critical entries so `pm activity` can
-    stay unstyled but still draw attention.
-
-    ``project_width`` left-pads the project key inside the brackets so
-    multi-row renders align even when project keys differ in length
-    (auto-fit; see :func:`compute_project_column_width` and #929). When
-    ``None`` the column is sized to the project key itself — the
-    single-row default that preserves the historical layout.
-    """
-    rel = format_relative_time(entry.timestamp, now=now)
-    project = _project_label(entry)
-    actor = entry.actor or "system"
-    verb = entry.verb or entry.kind
-    # #1033: distinguish create vs clear in the alert lifecycle so a
-    # quick scan tells the user which way the row is pointing. Only
-    # rewrite the verb when the emitter didn't supply a richer one
-    # (e.g. ``activity_summary(verb='alerted', ...)``) so structured
-    # events keep their authored verb intact.
-    if entry.kind == "alert" and verb in ("alert", entry.kind):
-        verb = "alert↑"
-    elif entry.kind == "alert.cleared" and verb in ("alert.cleared", entry.kind):
-        verb = "alert↓"
-    prefix = "!" if entry.severity == "critical" else " "
-    pin = "📌 " if entry.pinned else ""
-    width = max(project_width or len(project), len(project))
-    project_cell = f"{project:<{width}}"
-    return f"{prefix} {rel:>8}  [{project_cell}]  [{actor}]  {verb}: {pin}{entry.summary}"
-
-
-def render_entries_as_text(entries: Iterable[FeedEntry]) -> str:
-    """Render a list of FeedEntry rows as multi-line plain text.
-
-    Empty input renders a friendly placeholder so the cockpit panel
-    doesn't look broken on a brand-new install. The project column
-    auto-fits to the widest key in the batch so long keys like
-    ``blackjack-trainer`` aren't visually clipped (see #929).
-    """
-    materialised = list(entries)
-    if not materialised:
-        return (
-            "No activity yet.\n\n"
-            "Events accumulate as sessions start, tasks transition, "
-            "and heartbeats fire. Check back after the next sweep."
-        )
-    width = compute_project_column_width(materialised)
-    rows = [format_entry_row(e, project_width=width) for e in materialised]
-    return "\n".join(rows)
-
-
-# Re-exported from the core shared module (#1363). The plugin's ``FeedEntry``
-# dataclass satisfies ``ActivityFeedEntry`` structurally, so the renderer
-# accepts the existing concrete type unchanged.
-from pollypm.activity_feed_protocol import render_entry_detail  # noqa: F401, E402
-
-
-def render_activity_feed_text(config: Any, *, limit: int = 50) -> str:
-    """Render the feed as plain text for the cockpit's static pane path.
-
-    Picks up the projector from :func:`build_projector` — missing config
-    or missing state DB yields the same friendly placeholder as an
-    empty feed.
-    """
-    header = "Activity Feed"
-    projector = build_projector(config)
-    if projector is None:
-        return f"{header}\n\nNo state store configured — nothing to show yet."
-    try:
-        entries = projector.project(limit=limit)
-    except Exception:  # noqa: BLE001
-        logger.exception("activity_feed: projection failed for text render")
-        return f"{header}\n\nFailed to read activity events."
-    return f"{header}\n\n{render_entries_as_text(entries)}"
+# (and other peer plugins) can use these helpers without importing through
+# ``plugins_builtin`` (#1363). The plugin keeps the names on its own surface
+# for back-compat with existing imports + monkeypatch paths. The plugin's
+# ``FeedEntry`` dataclass satisfies ``ActivityFeedEntry`` structurally, so
+# the renderers accept the existing concrete type unchanged.
+from pollypm.activity_feed_protocol import (  # noqa: F401, E402
+    compute_project_column_width,
+    format_entry_row,
+    format_relative_time,
+    render_activity_feed_text,
+    render_entries_as_text,
+    render_entry_detail,
+)
+from pollypm.activity_feed_protocol import _project_label  # noqa: F401, E402
 
 
 # ---------------------------------------------------------------------------
